+++ src/view/explorer/ExplorerPane.ts
@@ -20,6 +20,7 @@ import { drafts, pluginSettings, selectedDraft } from "src/model/stores";
 import { insertScene } from "src/model/draft-utils";
 import NewDraftModal from "src/view/project-lifecycle/new-draft-modal";
 import { UndoManager } from "../undo/undo-manager";
+import { ignoreScene } from "./UseCases";
 
 export const VIEW_TYPE_LONGFORM_EXPLORER = "VIEW_TYPE_LONGFORM_EXPLORER";
 
@@ -290,76 +291,5 @@ export class ExplorerPane extends ItemView {
     if (this.explorerView) {
       this.explorerView.$destroy();
     }
-  }
-}
-
-export const getSelectedDraftWithIndex = () => {
-  const draft = get(selectedDraft) as MultipleSceneDraft;
-  const index = get(drafts).findIndex(
-    (d) => d.vaultPath === draft.vaultPath
-  );
-  return { index, draft }
-}
-
-export const addScene = (fileName: string) => {
-  const { index, draft } = getSelectedDraftWithIndex()
-  if (index >= 0 && draft.format === "scenes") {
-    drafts.update((d) => {
-      const targetDraft = d[index] as MultipleSceneDraft;
-      (d[index] as MultipleSceneDraft).scenes = [
-        ...targetDraft.scenes,
-        { title: fileName, indent: 0 },
-      ];
-      (d[index] as MultipleSceneDraft).unknownFiles =
-        targetDraft.unknownFiles.filter((f) => f !== fileName);
-      return d;
-    });
-  }
-}
-
-export const ignoreScene = (fileName: string) => {
-  const { index, draft } = getSelectedDraftWithIndex()
-  if (index >= 0 && draft.format === "scenes") {
-    drafts.update((d) => {
-      const targetDraft = d[index] as MultipleSceneDraft;
-      (d[index] as MultipleSceneDraft).scenes = targetDraft.scenes.filter(it => it.title != fileName);
-      (d[index] as MultipleSceneDraft).ignoredFiles = [
-        ...targetDraft.ignoredFiles,
-        fileName,
-      ];
-      (d[index] as MultipleSceneDraft).unknownFiles =
-        targetDraft.unknownFiles.filter((f) => f !== fileName);
-      return d;
-    });
-  }
-}
-
-export const addAll = () => {
-  const { index, draft } = getSelectedDraftWithIndex()
-  if (index >= 0 && draft.format === "scenes") {
-    drafts.update((d) => {
-      const targetDraft = d[index] as MultipleSceneDraft;
-      (d[index] as MultipleSceneDraft).scenes = [
-        ...targetDraft.scenes,
-        ...targetDraft.unknownFiles.map((f) => ({ title: f, indent: 0 })),
-      ];
-      (d[index] as MultipleSceneDraft).unknownFiles = [];
-      return d;
-    });
-  }
-}
-
-export const ignoreAll = () => {
-  const { index, draft } = getSelectedDraftWithIndex()
-  if (index >= 0 && draft.format === "scenes") {
-    drafts.update((d) => {
-      const targetDraft = d[index] as MultipleSceneDraft;
-      (d[index] as MultipleSceneDraft).ignoredFiles = [
-        ...targetDraft.ignoredFiles,
-        ...targetDraft.unknownFiles,
-      ];
-      (d[index] as MultipleSceneDraft).unknownFiles = [];
-      return d;
-    });
   }
 }
